Extract superclass from TestIndexController
